Moving out the CredCard component to a separate file.
